refactor: decouple register ui logic from business logic
