+++ app/src/main/AndroidManifest.xml
@@ -12,8 +12,13 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Planetze"
         tools:targetApi="31">
-
-        <!-- Main Activity -->
+        <activity
+            android:name=".MainActivity2"
+            android:exported="false" />
+        <activity
+            android:name=".auth.MainActivity"
+            android:exported="false"
+            android:theme="@style/Theme.Planetze" /> <!-- Main Activity -->
         <activity
             android:name=".auth.AuthActivity"
             android:exported="false" />
@@ -22,15 +27,13 @@
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
+
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
-        </activity>
-
-        <!-- Register Activity -->
+        </activity> <!-- Register Activity -->
         <activity
-            android:name="com.b07.planetze.Register"
+            android:name=".auth.RegisterFragment"
             android:exported="true" />
-
     </application>
 
 </manifest>
