+++ app/src/main/java/com/b07/planetze/auth/RegisterCallback.java
@@ -1,4 +1,7 @@
 package com.b07.planetze.auth;
 
+import androidx.annotation.NonNull;
+
 public interface RegisterCallback {
+    void register(@NonNull String email, @NonNull String password);
 }
