+++ app/src/main/java/com/b07/planetze/auth/RegisterFragment.java
@@ -1,4 +1,4 @@
-package com.b07.planetze;
+package com.b07.planetze.auth;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -6,26 +6,16 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
 import android.widget.EditText;
-import android.widget.Toast;
 
-import androidx.activity.EdgeToEdge;
-import androidx.annotation.NonNull;
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.core.graphics.Insets;
-import androidx.core.view.ViewCompat;
-import androidx.core.view.WindowInsetsCompat;
 import androidx.fragment.app.Fragment;
 
-import com.b07.planetze.auth.RegisterCallback;
-import com.google.android.gms.tasks.OnCompleteListener;
-import com.google.android.gms.tasks.Task;
-import com.google.firebase.auth.FirebaseAuth;
+import com.b07.planetze.R;
 
 /**
  * A user registration screen. <br>
  * Activities using this fragment must implement RegisterCallback.
  */
-public class Register extends Fragment {
+public class RegisterFragment extends Fragment {
 
     private Button btnCreate;
     private EditText editEmailRegister, editPasswordRegister;
@@ -53,5 +43,4 @@ public class Register extends Fragment {
 
         return view;
     }
-
 }
