+++ app/src/main/res/layout/activity_register.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Register">
+    tools:context=".auth.RegisterFragment">
 
     <Button
         android:id="@+id/btnCreateAccount"
