+++ app/src/main/res/navigation/nav_graph.xml
@@ -9,7 +9,7 @@
         android:id="@+id/FirstFragment"
         android:name="com.b07.planetze.auth.FirstFragment"
         android:label="@string/first_fragment_label"
-        tools:layout="@layout/fragment_first">
+        tools:layout="@layout/fragment_first" >
 
         <action
             android:id="@+id/action_FirstFragment_to_SecondFragment"
@@ -19,7 +19,7 @@
         android:id="@+id/SecondFragment"
         android:name="com.b07.planetze.auth.SecondFragment"
         android:label="@string/second_fragment_label"
-        tools:layout="@layout/fragment_second">
+        tools:layout="@layout/fragment_second" >
 
         <action
             android:id="@+id/action_SecondFragment_to_FirstFragment"
