+++ app/src/main/res/values-v23/themes.xml
@@ -1,9 +1,8 @@
 <resources xmlns:tools="http://schemas.android.com/tools">
-
-    <style name="Theme.Planetze" parent="Base.Theme.Planetze">
-        <!-- Transparent system bars for edge-to-edge. -->
-        <item name="android:navigationBarColor">@android:color/transparent</item>
-        <item name="android:statusBarColor">@android:color/transparent</item>
-        <item name="android:windowLightStatusBar">?attr/isLightTheme</item>
-    </style>
+  <style name="Theme.Planetze" parent="Base.Theme.Planetze">
+    <!-- Transparent system bars for edge-to-edge. -->
+    <item name="android:navigationBarColor">@android:color/transparent</item>
+    <item name="android:statusBarColor">@android:color/transparent</item>
+    <item name="android:windowLightStatusBar">?attr/isLightTheme</item>
+  </style>
 </resources>
